refactor: split the A*-attempt out of WaveguideRouter (file-size hard limit)
PR #758 grew WaveguideRouter.cs to 553 lines, tripping the 500-line
architecture guard on CI (the merge slipped through against a red
check). TryRouteAStar + AllowedRadiiIncluding move verbatim into the
partial WaveguideRouter.AStarAttempt.cs (375 + 191 lines) — no
behavioral change.
